requirements.txt
absl-py==1.0.0
astroid==2.4.2
astunparse==1.6.3
attrs==20.3.0
autopep8==1.5.4
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
cmake==3.22.1
colorama==0.4.4
comtypes==1.1.7
debugpy==1.5.1
decorator==4.4.2
dill==0.3.4
# dlib @ file:///C:/Users/bunty/OneDrive/Desktop/All%20Folders/Install-dlib-main/dlib-19.19.0-cp38-cp38-win_amd64.whl
docopt==0.6.2
face-recognition==1.3.0
face-recognition-models==0.3.0
flake8==3.8.4
flatbuffers==2.0
future==0.18.2
gast==0.4.0
geocoder==1.38.1
geographiclib==1.50
geopy==2.0.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.43.0
gTTS==2.2.3
h5py==3.6.0
httplib2==0.18.1
hurry==1.1
hurry.filesize==0.9
idna==2.10
importlib-metadata==4.10.0
isort==5.6.4
Js2Py==0.71
keras==2.7.0
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
libclang==12.0.0
loguru==0.6.0
lxml==4.6.2
Markdown==3.3.6
mccabe==0.6.1
more-itertools==8.6.0
MouseInfo==0.1.3
mpmath==1.2.1
multiprocess==0.70.12.2
mysql-connector-python==8.0.27
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pathos==0.2.8
Pillow==8.0.1
pipwin==0.5.1
playsound==1.3.0
pox==0.3.0
ppft==1.6.6.4
pprintpp==0.4.0
protobuf==3.14.0
psutil==5.7.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
# PyAudio @ file:///C:/Users/bunty/OneDrive/Desktop/All%20Folders/JARVIS-master/PyAudio-0.2.11-cp38-cp38-win_amd64.whl
PyAutoGUI==0.9.52
pycodestyle==2.6.0
PyExecJS==1.5.1
pyflakes==2.2.0
PyGetWindow==0.0.9
pyjokes==0.6.0
pyjsparser==2.7.1
pylint==2.6.0
PyMsgBox==1.0.9
pyparsing==3.0.6
PyPDF2==1.26.0
pyperclip==1.8.1
pypiwin32==223
PyPrind==2.11.3
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.0.1
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3
PyRect==0.1.4
PyScreeze==0.1.26
pySmartDL==1.3.4
python-dotenv==0.15.0
pyttsx3==2.90
PyTweening==1.0.3
pytz==2020.4
pytz-deprecation-shim==0.1.0.post0
pywhatkit==3.2
pywin32==300
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
ratelim==0.1.6
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.6
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
sympy==1.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
toml==0.10.2
translators==5.0.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==2.0.2
wikipedia==1.4.0
win32-setctime==1.1.0
wincertstore==0.2
winshell==0.6
wolframalpha==4.1.1
wrapt==1.12.1
xml-python==0.3.5
xmltodict==0.12.0
zipp==3.7.0
